requirements.txt
python-dotenv
spotipy
qrcode[pil]
pyzbar
Pillow
